Collapse all-scalar mutual tail-call SCCs at the shared monomorphize exit so native, wasm and the interpreter all run the loop
