+++ crates/almide-interp/interp-abstain-ledger.txt
@@ -103,7 +103,6 @@ mutable_global_bytes_arena  out-of-interp-scope capability: prim.handle
 mutable_global_bytes_copy_from  out-of-interp-scope capability: prim.handle
 mutable_global_repeat_writes  out-of-interp-scope capability: prim.handle
 mutual_append  out-of-interp-scope capability: prim.handle
-mutual_tail_recursion  interp fuel/recursion budget exhausted
 nan_canonical_bytes_write  out-of-interp-scope capability: prim.handle
 negative_offset_and_width_totality  out-of-interp-scope capability: prim.handle
 nested_hof_lambda_inference  out-of-interp-scope capability: prim.handle
