+++ crates/almide-optimize/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod optimize;
 pub mod mono;
+pub mod mutual_tco;
