+++ docs/contracts/contracts.toml
@@ -2237,7 +2237,7 @@ evidence  = [
 id        = "C-178"
 spec      = "ALS-M5"
 title     = "A mutual tail-recursion chain runs at unbounded depth on the wasm target"
-statement = "A cross-function tail-recursion chain (`is_even`/`is_odd`, an accumulator `ping`/`pong`) computes the same value on both targets at a depth (1,000,000) that overflows a plain-call chain on wasm (#864 — the legacy almide-codegen emitted return_call, but the port was missed when almide-mir became the default renderer: the chain compiled to plain `call`s and trapped `call stack exhausted`). The renderer now classifies a FUNCTION-TAIL `CallFn` — one whose result flows UNMODIFIED through only Else/EndIf value merges (plus its own Consume move-out marker) to the function's `ret`, with nothing else (no drops, no stores) on the path from the call to the exit — and emits `return_call`, transferring the frame; the dead merge plumbing after it is unreachable-but-valid wasm. The classification is render-side only (MIR ops, certificate, and caps counts are unchanged — the call op is the same op, differently encoded), and anything breaking the pattern keeps today's plain `call`. SELF tail recursion still takes the upstream TCO loop rewrite and is not affected. Native Rust emit is byte-identical (emit-diffs 0 across the 1001-file corpus); the wasm re-encode was suite-verified (all three spec legs plus flagship outputs byte-identical). The INDIRECT twin (2026-08-03): the same classification admits a function-tail `CallIndirect`, so a closure call in the reaches-ret-unmodified shape renders as `return_call_indirect` (the closure_tail_recursion fixture's `step`; the instruction encoding is machine-pinned by crates/almide-mir/tests/tail_call_indirect_test.rs). The depth claim stays honest: a cycle that ALLOCATES a closure per hop keeps one creator frame per round on BOTH targets — the creator's post-call drop correctly declines the transfer, a transferred frame would leak the closure — so the indirect fixture pins parity at a survivable depth plus the instruction gate, not an overflow-proving depth; an unbounded pure-indirect cycle needs a creator-free closure chain no spine-lowerable construction produces today."
+statement = "A cross-function tail-recursion chain (`is_even`/`is_odd`, an accumulator `ping`/`pong`) computes the same value on both targets at a depth (1,000,000) that overflows a plain-call chain on wasm (#864 — the legacy almide-codegen emitted return_call, but the port was missed when almide-mir became the default renderer: the chain compiled to plain `call`s and trapped `call stack exhausted`). The renderer now classifies a FUNCTION-TAIL `CallFn` — one whose result flows UNMODIFIED through only Else/EndIf value merges (plus its own Consume move-out marker) to the function's `ret`, with nothing else (no drops, no stores) on the path from the call to the exit — and emits `return_call`, transferring the frame; the dead merge plumbing after it is unreachable-but-valid wasm. The classification is render-side only (MIR ops, certificate, and caps counts are unchanged — the call op is the same op, differently encoded), and anything breaking the pattern keeps today's plain `call`. SELF tail recursion still takes the upstream TCO loop rewrite and is not affected. Native Rust emit is byte-identical (emit-diffs 0 across the 1001-file corpus); the wasm re-encode was suite-verified (all three spec legs plus flagship outputs byte-identical). The INDIRECT twin (2026-08-03): the same classification admits a function-tail `CallIndirect`, so a closure call in the reaches-ret-unmodified shape renders as `return_call_indirect` (the closure_tail_recursion fixture's `step`; the instruction encoding is machine-pinned by crates/almide-mir/tests/tail_call_indirect_test.rs). The depth claim stays honest: a cycle that ALLOCATES a closure per hop keeps one creator frame per round on BOTH targets — the creator's post-call drop correctly declines the transfer, a transferred frame would leak the closure — so the indirect fixture pins parity at a survivable depth plus the instruction gate, not an overflow-proving depth; an unbounded pure-indirect cycle needs a creator-free closure chain no spine-lowerable construction produces today. THE ALL-SCALAR SUBSET (#1043): a tail-call SCC whose members' params and shared return type are all Int/Float/Bool is collapsed BEFORE either renderer — in the shared IR at the monomorphize exit (almide-optimize's mutual_tco, a running-flag dispatcher loop plus thin per-member wrappers) — so this fixture's is_even/is_odd and ping/pong execute as the SAME loop on native, wasm AND almide-interp, and native termination no longer depends on the cargo profile's opt-level (the dependence #1043 objects to; pinned at true rustc opt-0 by native_v1_differential_test). `return_call`/`return_call_indirect` remain the wasm mechanism for every shape outside that gate — heap-typed params/returns and closure cycles, the closure_tail_recursion fixture among them — and the render-side classification is unchanged."
 since     = "0.37.0"
 status    = "active"
 evidence  = [
